requirements.txt
appdirs==1.4.4
argcomplete==1.12.1
astroid==2.4.2
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.6.2
distlib==0.3.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
googlemaps==4.4.2
httplib2==0.18.1
httpserver==1.1.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nox==2020.8.22
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
pandas @ file:///C:/ci/pandas_1602083338010/work
protobuf==3.14.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
responses==0.12.1
rsa==4.6
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.1.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
